requirements.txt
# === Core API ===
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# === ASR ===
openai-whisper==20231117
ffmpeg-python==0.2.0

# === Intent Classification ===
transformers==4.41.0
torch==2.3.0
datasets==2.19.0
scikit-learn==1.4.2
accelerate==0.30.0

# === TTS ===
gTTS==2.5.1
pydub==0.25.1

# === Evaluation & Metrics ===
jiwer==3.0.4
seaborn==0.13.2
matplotlib==3.9.0

# === Utilities ===
pyyaml==6.0.1
loguru==0.7.2
numpy==1.26.4
tqdm==4.66.4

# === Dev / Testing ===
pytest==8.2.0
httpx==0.27.0